Rename Monday asset system to TH.e-1RL_MONDAY.Ng
