+++ apps/api/the_one_real_monday.py
@@ -1,7 +1,7 @@
-"""API endpoints for MONDAY asset generation."""
+"""API endpoints for TH.e-1RL_MONDAY.Ng asset generation."""
 import json
 from pathlib import Path
-from tools.monday_generator import generate_assets
+from tools.the_one_real_monday_generator import generate_assets
 
 
 def generate():
