+++ frontend/pages/theOneRealMonday.jsx
@@ -18,21 +18,21 @@ function AssetGallery({ assets }) {
   );
 }
 
-export default function MondayAssetTool() {
+export default function TheOneRealMondayTool() {
   const [assets, setAssets] = useState(null);
   const [loading, setLoading] = useState(false);
 
   async function handleGenerate() {
     setLoading(true);
-    await fetch('/api/monday/generate', { method: 'POST' });
-    const data = await fetch('/api/monday/assets_map').then((r) => r.json());
+    await fetch('/api/the_one_real_monday/generate', { method: 'POST' });
+    const data = await fetch('/api/the_one_real_monday/assets_map').then((r) => r.json());
     setAssets(data);
     setLoading(false);
   }
 
   return (
-    <div className="monday-tool">
-      <h1>MONDAY Asset Generator</h1>
+    <div className="one-real-monday-tool">
+      <h1>TH.e-1RL_MONDAY.Ng Asset Generator</h1>
       <button onClick={handleGenerate} disabled={loading}>
         {loading ? 'Generating...' : 'Generate Assets'}
       </button>
